requirements.txt
beanie==1.16.8
quart==0.18.4
python-dotenv==0.20.0
